ci(examples): add 'journey test' examples
