+++ .github/workflows/examples.yaml
@@ -1,0 +1,27 @@
+name: Run examples
+
+on:
+  push:
+    branches:
+      - v6
+
+jobs:
+  examples:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        EXAMPLE:
+          - "basic"
+    env:
+      # TODO(dyma): update GH secrets names
+      WEAVIATE_HOST: ${{ secrets.WEAVIATE_HOST }}
+      WEAVIATE_API_KEY: ${{ secrets.WEAVIATE_API_KEY }}
+    steps:
+      - uses: actions/checkout@v6
+      - uses: actions/setup-go@v6
+        with:
+          go-version: "1.24"
+          cache: true
+      - name: example/${{ matrix.EXAMPLE }}
+        run: go run ./example/${{ matrix.EXAMPLE }}
